requirements-dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --extra dev pyproject.toml -o requirements-dev.txt
aiohttp==3.9.5
    # via
    #   aiosnow (pyproject.toml)
    #   pytest-aiohttp
aiosignal==1.3.1
    # via aiohttp
alabaster==0.7.16
    # via sphinx
anyio==4.4.0
    # via httpcore
attrs==23.2.0
    # via aiohttp
babel==2.15.0
    # via sphinx
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via codecov-cli
codecov-cli==0.6.0
    # via aiosnow (pyproject.toml)
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   sphinx
coverage==7.5.3
    # via pytest-cov
cryptography==42.0.7
    # via pyjwt
docutils==0.21.2
    # via
    #   aiosnow (pyproject.toml)
    #   readme-renderer
    #   sphinx
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
httpcore==0.16.3
    # via httpx
httpx==0.23.3
    # via codecov-cli
idna==3.7
    # via
    #   anyio
    #   requests
    #   rfc3986
    #   yarl
ijson==3.2.3
    # via codecov-cli
imagesize==1.4.1
    # via sphinx
importlib-metadata==7.1.0
    # via twine
iniconfig==2.0.0
    # via pytest
jaraco-classes==3.4.0
    # via keyring
jaraco-context==5.3.0
    # via keyring
jaraco-functools==4.0.1
    # via keyring
jinja2==3.1.4
    # via sphinx
keyring==25.2.1
    # via twine
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
marshmallow==3.21.2
    # via aiosnow (pyproject.toml)
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.2.0
    # via
    #   jaraco-classes
    #   jaraco-functools
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy==1.10.0
    # via aiosnow (pyproject.toml)
mypy-extensions==1.0.0
    # via mypy
nh3==0.2.17
    # via readme-renderer
packaging==24.0
    # via
    #   marshmallow
    #   pytest
    #   sphinx
pkginfo==1.11.0
    # via twine
pluggy==1.5.0
    # via pytest
pockets==0.9.1
    # via sphinxcontrib-napoleon
pycparser==2.22
    # via cffi
pygments==2.18.0
    # via
    #   readme-renderer
    #   rich
    #   sphinx
pyjwt==2.8.0
    # via aiosnow (pyproject.toml)
pytest==8.2.1
    # via
    #   aiosnow (pyproject.toml)
    #   pytest-aiohttp
    #   pytest-asyncio
    #   pytest-cov
pytest-aiohttp==1.0.5
    # via aiosnow (pyproject.toml)
pytest-asyncio==0.23.7
    # via
    #   aiosnow (pyproject.toml)
    #   pytest-aiohttp
pytest-cov==5.0.0
    # via aiosnow (pyproject.toml)
pywin32-ctypes==0.2.2
    # via keyring
pyyaml==6.0.1
    # via codecov-cli
readme-renderer==43.0
    # via twine
regex==2024.5.15
    # via codecov-cli
requests==2.32.3
    # via
    #   requests-toolbelt
    #   responses
    #   sphinx
    #   twine
requests-toolbelt==1.0.0
    # via twine
responses==0.21.0
    # via codecov-cli
rfc3986==1.5.0
    # via
    #   httpx
    #   twine
rich==13.7.1
    # via twine
setuptools==70.0.0
    # via tree-sitter
six==1.16.0
    # via
    #   pockets
    #   sphinxcontrib-napoleon
sniffio==1.3.1
    # via
    #   anyio
    #   httpcore
    #   httpx
snowballstemmer==2.2.0
    # via sphinx
sphinx==7.3.7
    # via aiosnow (pyproject.toml)
sphinxcontrib-applehelp==1.0.8
    # via sphinx
sphinxcontrib-devhelp==1.0.6
    # via sphinx
sphinxcontrib-htmlhelp==2.0.5
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-napoleon==0.7
    # via aiosnow (pyproject.toml)
sphinxcontrib-qthelp==1.0.7
    # via sphinx
sphinxcontrib-serializinghtml==1.1.10
    # via sphinx
test-results-parser==0.1.0
    # via codecov-cli
tree-sitter==0.20.4
    # via codecov-cli
twine==5.1.0
    # via aiosnow (pyproject.toml)
types-pytz==2024.1.0.20240417
    # via aiosnow (pyproject.toml)
typing-extensions==4.12.1
    # via mypy
urllib3==2.2.1
    # via
    #   requests
    #   responses
    #   twine
yarl==1.9.4
    # via aiohttp
zipp==3.19.1
    # via importlib-metadata
